requirements.txt
python-telegram-bot>=13.0
web3>=5.0.0
python-dotenv>=0.19.0
requests>=2.25.0
schedule>=1.1.0
eth-account>=0.5.6
eth-utils>=1.10.0
eth-typing>=2.2.2
pycryptodome>=3.10.1
aiohttp>=3.7.4
cryptography>=3.4.8
pytz>=2021.1
